@@ -8,33 +8,32 @@ version = "0.3.0"
 description = "Cheat-sheets for shortcuts & commands at your fingertips."
 keywords = ["shortcuts", "keybindings", "hints", "helper", "cheatsheet"]
 readme = "README.md"
-requires-python = ">=3.7"
-license = "MIT"
+requires-python = ">=3.11"
 authors = [{ name = "dynobo", email = "dynobo@mailbox.org" }]
 classifiers = [
   "Development Status :: 4 - Beta",
   "License :: OSI Approved :: MIT License",
   "Programming Language :: Python",
-  "Programming Language :: Python :: 3.7",
-  "Programming Language :: Python :: 3.8",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: Implementation :: CPython",
   "Programming Language :: Python :: Implementation :: PyPy",
   "Topic :: Utilities",
   "Intended Audience :: End Users/Desktop",
   "Operating System :: POSIX :: Linux",
 ]
-dependencies = ["PyGObject>=3.42.2", "PyYAML>=6.0"]
+dependencies = ["PyGObject>=3.42.2"]
 
 [project.urls]
 Documentation = "https://github.com/dynobo/keyhint#readme"
 Issues = "https://github.com/dynobo/keyhint/issues"
 Source = "https://github.com/dynobo/keyhint"
 
+[project.scripts]
+keyhint = "keyhint.app:main"
+
 [[tool.hatch.envs.all.matrix]]
-python = ["3.7", "3.8", "3.9", "3.10", "3.11"]
+python = ["3.11"]
 
 [tool.hatch.envs.default]
 dependencies = [
@@ -43,7 +42,7 @@ dependencies = [
   "pytest-cov",
   "pre-commit",
   "coveralls",
-  "types-PyYAML",
+  "types-toml",
   "tbump",
   "ruff",
   "pip-audit",
@@ -69,8 +68,9 @@ fmt = [
 check = ["style", "typing", "test", "security"]
 pre-commit = ["fmt", "check"]
 version = "tbump {args:current-version}"
+
 [tool.ruff]
-target-version = "py39"
+target-version = "py311"
 line-length = 88
 exclude = [".venv"]
 select = [
