@@ -5,9 +5,9 @@ from pathlib import Path
 from keyhint import utils
 
 
-def test_load_default_hints():
-    """Test loading of yaml files shipped with package."""
-    hints = utils.load_default_hints()
-    yaml_files = (Path(__file__).parent.parent / "keyhint" / "config").glob("*.yaml")
+def test_load_default_sheets():
+    """Test loading of toml files shipped with package."""
+    sheets = utils.load_default_sheets()
+    toml_files = (Path(__file__).parent.parent / "keyhint" / "config").glob("*.toml")
 
-    assert len(hints) == len(list(yaml_files))
+    assert len(sheets) == len(list(toml_files))
